move geosjon to tracks conversion util
